2002-09-23  Michal Ludvig  <mludvig@suse.cz> 	Merged x86-64 port by Andi Kleen <ak@suse.de> 	and Michal Ludvig <mludvig@suse.cz> 	* Makefile.in: New target 'headers'. Failure ignored. 	* acconfig.h: New defines for x86-64. 	* configure.in: Ditto. 	* defs.h: Ditto. 	* file.c: Ditto. 	* signal.c: Ditto. 	* process.c: Added support for x86-64. 	* util.c: Ditto. 	* syscall.c: Ditto + added automatic personality switching. 	* linux/syscall.h: Ditto. 	* linux/x86_64: New directory. 	* linux/x86_64/Makefile.in: New file. 	* linux/x86_64/gentab.pl: Ditto. 	* linux/x86_64/i386-headers.diff: Ditto. 	* linux/x86_64/makeheaders.sh: Ditto. 	* linux/x86_64/syscallent.h: Ditto. 	* mem.c (print_mmap): Always print arg[4] as int.

+++ mem.c
@@ -185,8 +185,8 @@ long *u_arg;
 #else
 		printflags(mmap_flags, u_arg[3]);
 #endif
-		/* fd */
-		tprintf(", %ld, ", u_arg[4]);
+		/* fd (is always int, not long) */
+		tprintf(", %d, ", (int)u_arg[4]);
 		/* offset */
 		tprintf("%#lx", u_arg[5]);
 	}
